changed tests creation and view so the logged in teacher will only be able to create tests for his courses and view tests for his courses. didnt run it yet.

+++ TestUs/src/main/java/com/springbootquickstart/TestUs/repository/TeacherRepository.java
@@ -8,9 +8,8 @@ import org.springframework.stereotype.Repository;
 import java.util.Optional;
 
 @Repository
-public interface TeacherRepository extends JpaRepository<Teacher, Long> {
+public interface TeacherRepository extends JpaRepository<Teacher, Integer> {
 
     Optional<Teacher> findByMyUser(MyUser myUser);
-
 
 }
